changing the format of the ID for steps to be returned. Still need to add an ID for the stepitself and fix the case of no steps being returned. It is currently returning a 500 whatg I dont want.
